libui: use uint64_t in dump and getTotalSize to avoid overflow issues
Fixes add-overflow errors caught by ubsan when running AHardwareBuffer
VK tests.

Test: dEQP-VK.api.external.memory.android_hardware_buffer.image_formats#r16g16b16a16_sfloat

+++ libs/ui/include/ui/GraphicBufferAllocator.h
@@ -49,7 +49,7 @@ public:
 
     status_t free(buffer_handle_t handle);
 
-    size_t getTotalSize() const;
+    uint64_t getTotalSize() const;
 
     void dump(std::string& res) const;
     static void dumpToSystemLog();
